refactor(tools): extract path-validation helpers from file-tools.ts into file-utils.ts
Round 6 Candidate #2 — extract path-validation helpers (~120 lines) from file-tools.ts (449 lines) into file-utils.ts.

- file-utils.ts (new): handleFileError, handleDirError, isSensitiveFile, validatePath, validateEditString, SENSITIVE_PATTERNS
- file-tools.ts (modified, -130 lines): imports from file-utils.ts, re-exports for backward compat
